fix/ added guard test on uid nullity before sending notification
Took 1 hour 4 minutes
