+++ app/src/main/java/com/example/event_app/repository/NotificationRepository.kt
@@ -10,7 +10,7 @@ import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import com.example.event_app.R
 import com.example.event_app.model.Message
-import com.example.event_app.ui.activity.MainActivity
+import com.example.event_app.ui.activity.LoginActivity
 import com.google.firebase.database.FirebaseDatabase
 import timber.log.Timber
 
@@ -51,7 +51,7 @@ class NotificationRepository(private val context: Context) {
         context: Context) {
 
         if (currentUserId == dataMessage) {
-            val intent = Intent(context, MainActivity::class.java)
+            val intent = Intent(context, LoginActivity::class.java)
             intent.putExtra("title", dataTitle)
             intent.putExtra("message", dataMessage)
             intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
