+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Event-App</string>
+    <string name="app_name">Event Moi Ça</string>
 
     <!-- Title -->
     <string name="title_home">Évènements</string>
